test_size should be a float

+++ preprocessing.py
@@ -15,7 +15,7 @@ from numpy.lib.stride_tricks import sliding_window_view
 
 parser = argparse.ArgumentParser()
 parser.add_argument('--dir', type=str, default='')
-parser.add_argument('--test_size', type=int, default=0)
+parser.add_argument('--test_size', type=float, default=0)
 parser.add_argument('--crop_size', type=int, nargs="+", default=(1, 64, 64))
 parser.add_argument('--strides', type=int, nargs="+", default=None)
 parser.add_argument('--save_name', type=str, default="processed")
